api/pyproject.toml
[tool.poetry]
name = "ai-summit-lab2-api"
version = "0.1.0"
description = ""
authors = ["Fu-Ming Tsai <716585+sary357@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pydantic = "^2.10.1"
poetry-plugin-dotenv = "^2.5.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
